+++ misata/__init__.py
@@ -24,7 +24,7 @@ Quickstart::
     tables = misata.generate_from_schema(gen.generate_from_story("A fintech fraud dataset"))
 """
 
-__version__ = "0.8.3.1"
+__version__ = "0.8.4"
 __author__ = "Muhammed Rasin"
 
 from typing import Any, Dict, Optional
@@ -554,6 +554,7 @@ from misata.schema import (
     Column,
     Constraint,
     NoiseConfig,
+    GroupShares,
     OutcomeCurve,
     RateCurve,
     RealismConfig,
@@ -716,6 +717,7 @@ __all__ = [
     "Column",
     "Constraint",
     "NoiseConfig",
+    "GroupShares",
     "OutcomeCurve",
     "RateCurve",
     "RealismConfig",